requirements.txt
langchain_openai
langchain_core
streamlit~=1.35.0
python-dotenv~=1.0.1
langchain_community
langchain_google_genai
langserve~=0.2.1
fastapi~=0.111.0
uvicorn~=0.30.1
sse_starlette
requests~=2.32.3
langchain~=0.2.3